Add the box-sizing property to all html elements
Ensures an equal width for all the box elements used in the web pages.

Create div sections in the chapati recipe pages

These are two allow for proper positioning of elements within the two
colums of the page.

Center align text and images

This create a neat and organized workflow of the various element objects